+++ aoc/coord.hpp
@@ -7,7 +7,7 @@
 namespace coord_util
 {
 
-struct coord { size_t x = 0; size_t y = 0; };
+struct coord { ptrdiff_t x = 0; ptrdiff_t y = 0; };
 
 bool operator< (const coord& a, const coord& b);
 std::ostream& operator<< (std::ostream& os, const coord& c);
@@ -20,8 +20,8 @@ bool operator==(const coord& a, const coord& b);
 
 struct coord_incrementer
 {
-    size_t x_limit = 0;
-    void operator()(coord& c, int by = 1) const;
+    ptrdiff_t x_limit = 0;
+    void operator()(coord& c, ptrdiff_t by = 1) const;
 };
 
 } // end namaspace coord_util
